use common gps time function
